LayoutTests/platform/chromium/media/video-frame-size-change-expected.txt

Test that a video element scales decoded frames to match the initial size as opposed to changing the size of the element.

canplay
canplay
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
EXPECTED (320 == '320') OK
EXPECTED (180 == '180') OK
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
playing
playing
pause
pause
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
EXPECTED (320 == '320') OK
EXPECTED (180 == '180') OK
EXPECTED (1920 == '1920') OK
EXPECTED (1080 == '1080') OK
END OF TEST
